getById method in User Repository connected to psql database and returns User object

+++ src/main/java/com/carDealership/model/User.java
@@ -28,7 +28,7 @@ public class User {
         return id;
     };
 
-    public User setId(int id) {
+    public User setId(long id) {
         this.id = id;
         return this;
     };
